refactor: drop unused dataset domain and organic question upsert
